% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{chunk}
\alias{chunk}
\title{Chunk a datatable to disk for parallel operations}
\usage{
chunk(dt, col, chunks = detectCores())
}
\arguments{
\item{dt}{A data table.}

\item{col}{Column name to chunk by.}

\item{chunks}{Number of chunks.}
}
\description{
Chunk a datatable to disk for parallel operations
}
