% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{coerce_dt}
\alias{coerce_dt}
\title{Recursively coerce a list to a flat data table, correctly naming columns.}
\usage{
coerce_dt(l, ix = NULL, names = TRUE)
}
\arguments{
\item{l}{List to coerce.}

\item{ix}{Numeric vector. Vector of nested indices to extract recursively.}

\item{names}{Logical. Preserve names?}
}
\value{
Data table with non-coercible objects elided.
}
\description{
Recursively coerce a list to a flat data table, correctly naming columns.
}
