% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{fbind}
\alias{fbind}
\title{Sequentially load and row bind data tables.}
\usage{
fbind(files)
}
\arguments{
\item{files}{Character vector. Files to load and row bind.}
}
\value{
A data table.
}
\description{
\code{fbind} sequentially loads data tables from disk and binds the data tables together by row names. The purpose of this method is to support parallelized data table operations, bypassing the 1.5 Gb return vector limit of the \code{lapply} family of functions. Both \code{.RDS} and plaint text tables are supported.
}
