% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%exclude\%}
\alias{\%exclude\%}
\title{Convenience inflix operator to return vector elements excluding regxpr.}
\usage{
vector \%exclude\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\description{
Convenience inflix operator to return vector elements excluding regxpr.
}
