% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%include\%}
\alias{\%include\%}
\title{Convenience inflix operator to return vector elements matching regxpr.}
\usage{
vector \%include\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\description{
Convenience inflix operator to return vector elements matching regxpr.
}
