% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%includef\%}
\alias{\%includef\%}
\title{Convenience inflix operator to return vector elements matching fixed pattern.}
\usage{
vector \%includef\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\description{
Convenience inflix operator to return vector elements matching fixed pattern.
}
