% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%likef\%}
\alias{\%likef\%}
\title{Convenience inflix operator to return logical vector of elements matching fixed pattern.}
\usage{
vector \%likef\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\description{
Convenience inflix operator to return logical vector of elements matching fixed pattern.
}
