% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%with\%}
\alias{\%with\%}
\title{Get dt columns by regexpr.}
\usage{
dt \%with\% cols
}
\arguments{
\item{dt}{A data table.}

\item{cols}{Vector of column patterns to include.}
}
\description{
Get dt columns by regexpr.
}
