% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{\%without\%}
\alias{\%without\%}
\title{Convenience inflix operator to remove columns from a data.table by reference.}
\usage{
dt \%without\% cols
}
\arguments{
\item{dt}{A data.table.}

\item{cols}{Vector of columns patterns to remove.}
}
\value{
A data.table subset of \code{dt} with column names not matching \code{cols}.
}
\description{
Convenience inflix operator to remove columns from a data.table by reference.
}
\examples{
library(magrittr)
library(data.table)
library(dt.inflix)

  dt <- data.table::data.table(a = c(1, 1, 1),
                               b = c(2, 2, 3))

  dto <- dt \%without\% "b"
  print(dto)
}
