% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%withoutrows\%}
\alias{\%withoutrows\%}
\title{Remove dt rows by reference.}
\usage{
dt \%withoutrows\% rows
}
\arguments{
\item{dt}{A data table.}

\item{rows}{Row indicies to delete.}
}
\description{
Remove dt rows by reference.
}
