% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{allduplicated}
\alias{allduplicated}
\title{Return all duplicated rows of a data.table.}
\usage{
allduplicated(dt, by = NULL)
}
\arguments{
\item{dt}{Data table.}

\item{by}{Character vector. Names of columns to test for duplicates.}
}
\value{
A data.table of duplicated rows from \code{dt}.
}
\description{
Return all duplicated rows of a data.table.
}
\examples{
library(magrittr)
library(data.table)
library(dt.inflix)

  dt <- data.table::data.table(a = c(1, 1, 1),
                               b = c(2, 2, 3))

  dt \%>\% allduplicated
}
