% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{allduplicated}
\alias{allduplicated}
\title{Return all duplicated rows of a data table.}
\usage{
allduplicated(dt, by = NULL)
}
\arguments{
\item{dt}{Data table.}
}
\description{
Return all duplicated rows of a data table.
}
