% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{chunk}
\alias{chunk}
\title{Chunk a data table to disk for parallel operations}
\usage{
chunk(dt, chunks = parallel::detectCores(), col.names = TRUE,
  sep = "\\t")
}
\arguments{
\item{dt}{A data.table.}

\item{chunks}{Number of chunks.}

\item{col.names}{Logical. Write column names?}

\item{sep}{Character. Column separator.}
}
\description{
Write a data.table to disk using \code{data.table::fwrite} in \code{chunks} segments. If is often faster to write out large tabular data to disk then read in parallel vs. \code{foreach} or \code{mclapply}
}
\examples{
\dontrun{
library(magrittr)
library(data.table)
library(dt.inflix)

  dt <- data.table::data.table(a = c(1:1E7),
                               b = LETTERS[1:20])

  dt \%>\% chunk
}
}
