% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{chunk}
\alias{chunk}
\title{Data.table::split but with aggregate "chunks".}
\usage{
chunk(dt, fun, by = ".id", cl = parallel::detectCores())
}
\arguments{
\item{dt}{A data.table.}

\item{fun}{Monadic data table function to use on chunks. Function must return a data table of columns to add and `.id`, and index created by `chunk`.

Example function:
```r
 .fn <- function(dt){
   print("start")
    dt[, uuid := full_name %>%
       stringr::str_extract("[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}")]

   return(dt[, .SD, .SDcols = c("uuid", ".chunk_id")])

 }
```r}

\item{by}{Column to chunk by.}

\item{cl}{Number of chunks.}
}
\value{
`cl` data.table chunks with `by` split exclusively. The data table returned has new columns called `.id`, the default by parameter of non was specified, and .vec_chunk, the assignment in `cl`.
}
\description{
`chunk` performs `data.table::split` for a column and then uses `fun` on each chunk. The purpose of this function is to parallelize over `by` more efficiently when `by` >> `cl`.
}
