% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%excludef\%}
\alias{\%excludef\%}
\title{Convenience inflix operator to return vector elements excluding fixed pattern.}
\usage{
vector \%excludef\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\description{
Convenience inflix operator to return vector elements excluding fixed pattern.
}
