% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{\%includef\%}
\alias{\%includef\%}
\title{Convenience inflix operator to return vector elements matching a fixed pattern.}
\usage{
vector \%includef\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\value{
A vector of elements in \code{vector} matching \code{pattern}.
}
\description{
Convenience inflix operator to return vector elements matching a fixed pattern.
}
\examples{
library(magrittr)
library(dt.inflix)

  v <- c("Apple", "apple", "APPLE")

  v \%include\% "Ap"
}
