% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{\%likef\%}
\alias{\%likef\%}
\title{Convenience inflix operator to return logical vector of elements matching a fixed pattern.}
\usage{
vector \%likef\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\value{
A logical vector the same length as \code{vector} of elements matching \code{pattern}.
}
\description{
Convenience inflix operator to return logical vector of elements matching a fixed pattern.
}
