% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{\%with\%}
\alias{\%with\%}
\title{Convenience inflix operator to return a data.table of columns matching a regular expression.}
\usage{
dt \%with\% cols
}
\arguments{
\item{dt}{A data.table.}

\item{cols}{Vector of column patterns to include.}
}
\value{
A data.table subset of \code{dt} with column names matching \code{cols}.
}
\description{
Convenience inflix operator to return a data.table of columns matching a regular expression.
}
