% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%without\%}
\alias{\%without\%}
\title{Remove columns from a data.table by reference}
\usage{
dt \%without\% cols
}
\arguments{
\item{dt}{A data table.}

\item{cols}{Vector of columns patterns to remove.}
}
\description{
Remove columns from a data.table by reference
}
