% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{\%withoutrows\%}
\alias{\%withoutrows\%}
\title{Convenience inflix operator to remove data.table rows.}
\usage{
dt \%withoutrows\% rows
}
\arguments{
\item{dt}{A data.table.}

\item{rows}{Row indicies to delete.}
}
\value{
A data.table subset of \code{dt} with rows from index \code{rows}.
}
\description{
Convenience inflix operator to remove data.table rows.
}
