% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.inflix.R
\name{withoutna}
\alias{withoutna}
\title{Convenience inflix operator to remove all(NA) or all(NULL) columns from a data.table by reference.}
\usage{
withoutna(dt)
}
\arguments{
\item{dt}{A data.table.}
}
\value{
A data.table subset of \code{dt} with columns of all \code{NA} or \code{NULL} removed.
}
\description{
Remove columns from a data table if all values are NA or NULL; useful when subsetting.
}
\examples{
library(magrittr)
library(data.table)
library(dt.inflix)

  dt <- data.table::data.table(a = c(1, 1, 1),
                               b = c(NA, NA, NA))

  dt \%>\% withoutna
  print(dt)
}
