% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.inflix.operators.R
\name{withoutna}
\alias{withoutna}
\title{Remove NA/NULL columns from a data table by reference}
\usage{
withoutna(dt)
}
\arguments{
\item{dt}{A data table.}
}
\description{
Remove columns from a data table if all values are NA or NULL; useful when subsetting.
}
